fix: remove android tests for CI
